requirements.txt
streamlit
numpy
collections
pillow
glob
os
tensorflow